Hide RootNodeOwnerImpl
